+++ be/src/main/java/com/jjikmuk/sikdorak/auth/controller/OAuthController.java
@@ -1,12 +1,14 @@
 package com.jjikmuk.sikdorak.auth.controller;
 
-import com.jjikmuk.sikdorak.auth.controller.response.JwtTokenResponse;
+import com.jjikmuk.sikdorak.auth.domain.JwtTokenPair;
 import com.jjikmuk.sikdorak.auth.controller.response.SikdorakAccessToken;
 import com.jjikmuk.sikdorak.auth.service.OAuthService;
 import com.jjikmuk.sikdorak.common.ResponseCodeAndMessages;
 import com.jjikmuk.sikdorak.common.response.CommonResponseEntity;
 import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
@@ -28,20 +30,27 @@ public class OAuthController {
     }
 
     @GetMapping("/api/oauth/callback")
-    public CommonResponseEntity<JwtTokenResponse> loginCallback(@RequestParam String code) {
+    public CommonResponseEntity<SikdorakAccessToken> loginCallback(@RequestParam String code, HttpServletResponse response) {
+        JwtTokenPair jwtTokenPair = oAuthService.login(code);
+        String refreshToken = jwtTokenPair.getRefreshToken();
+        setCookie(response, refreshToken);
+
         return new CommonResponseEntity<>(ResponseCodeAndMessages.LOGIN_SUCCESS,
-            oAuthService.login(code), HttpStatus.OK);
+            new SikdorakAccessToken(jwtTokenPair.getAccessToken()), HttpStatus.OK);
     }
 
     @GetMapping("/api/oauth/refresh")
-    public CommonResponseEntity<SikdorakAccessToken> updateAccessToken(
-        @CookieValue("refreshToken") Cookie cookie) {
-
+    public CommonResponseEntity<SikdorakAccessToken> updateAccessToken(@CookieValue("refreshToken") Cookie cookie) {
         String refreshToken = cookie.getValue();
 
         return new CommonResponseEntity<>(ResponseCodeAndMessages.UPDATE_ACCESS_TOKEN_SUCCESS,
             oAuthService.updateAccessToken(refreshToken),
             HttpStatus.OK);
+    }
 
+    private static void setCookie(HttpServletResponse response, String refreshToken) {
+        Cookie cookie = new Cookie("refreshToken", refreshToken);
+        cookie.setHttpOnly(true);
+        response.addCookie(cookie);
     }
 }
