+++ be/src/main/java/com/jjikmuk/sikdorak/auth/domain/JwtTokenPair.java
@@ -1,16 +1,16 @@
-package com.jjikmuk.sikdorak.auth.controller.response;
+package com.jjikmuk.sikdorak.auth.domain;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class JwtTokenResponse {
+public class JwtTokenPair {
 
     private String accessToken;
     private String refreshToken;
 
-    public JwtTokenResponse(String accessToken, String refreshToken) {
+    public JwtTokenPair(String accessToken, String refreshToken) {
         this.accessToken = accessToken;
         this.refreshToken = refreshToken;
     }
